Refactor components folders
